+++ src/main/__tests__/index-test.js
@@ -4,7 +4,7 @@ require('../index');
 const { Menu } = require('electron');
 
 // Don't start the server for these tests
-jest.mock('../worker/serverManager', () => ({
+jest.mock('../server/ServerFactory', () => ({
   serverEvents: {},
   createServer: jest.fn().mockResolvedValue({ connectionInfo: {}, on: jest.fn() }),
 }));
