+++ src/main/__tests__/settings-test.js
