+++ src/main/server/__tests__/ServerFactory-test.js
@@ -1,7 +1,7 @@
 /* eslint-env jest, jasmine */
 
 const path = require('path');
-const { createServer } = require('../serverManager');
+const { createServer } = require('../ServerFactory');
 
 const mockServerMethods = {
   close: jest.fn(),
