+++ src/main/server/devices/PairingCodeFactory.js
