+++ src/main/server/devices/PairingRequestService.js
@@ -5,8 +5,8 @@ const uuidv4 = require('uuid/v4');
 const libsodium = require('libsodium-wrappers');
 const logger = require('electron-log');
 
-const PairingCodeFactory = require('./pairingCodeFactory');
-const { RequestError } = require('../errors/RequestError');
+const PairingCodeFactory = require('./PairingCodeFactory');
+const { RequestError } = require('../../errors/RequestError');
 
 const DeviceRequestTTLSeconds = 5 * 60;
